Refactored FlyoutPopupPanel to extract placement logic
