+++ src/Uno.UI/UI/Xaml/Controls/Popup/PopupPanel.cs
@@ -29,7 +29,7 @@ namespace Windows.UI.Xaml.Controls
 
 		public PopupPanel(Popup popup)
 		{
-			Popup = popup;
+			Popup = popup ?? throw new ArgumentNullException(nameof(popup));
 			Visibility = Visibility.Collapsed;
 			Background = SolidColorBrushHelper.Transparent;
 		}
